requirements.txt
# Shadow Supply Chain Hunter - GraphRAG Pipeline
# Core dependencies for SEC filing analysis and graph-based reasoning

# LlamaIndex ecosystem - Using the modular architecture for better control
llama-index>=0.10.0
llama-index-core>=0.10.0
llama-index-graph-stores-neo4j>=0.2.0
llama-index-llms-openai>=0.1.0
llama-index-embeddings-openai>=0.1.0
llama-index-readers-file>=0.1.0

# Neo4j - Graph database driver for direct queries when needed
neo4j>=5.0.0

# SEC Data Ingestion
sec-edgar-downloader>=5.0.0

# Environment & Config
python-dotenv>=1.0.0

# Data Processing - Pandas for any tabular transformations
pandas>=2.0.0

# Optional: For local LLM support (Ollama)
# llama-index-llms-ollama>=0.1.0

# Utilities
tqdm>=4.65.0
